ocr-service/requirements.txt
easyocr>=1.7.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
Pillow>=10.4.0
